+++ app/assets/javascripts/Components/Modals/submission_file_upload_modal.jsx
@@ -101,12 +101,17 @@ class SubmissionFileUploadModal extends React.Component {
     return fileRenameInput;
   };
 
+  clearState = () => {
+    this.setState({newFiles: []});
+  };
+
   render() {
     return (
       <Modal
         className="react-modal"
         isOpen={this.props.isOpen}
         onRequestClose={this.props.onRequestClose}
+        onAfterClose={this.clearState}
       >
         <h2>{I18n.t("upload")}</h2>
         <form onSubmit={this.onSubmit}>
@@ -136,7 +141,11 @@ class SubmissionFileUploadModal extends React.Component {
               {this.fileRenameInputBox()}
             </label>
             <div className={"modal-container"}>
-              <input type="submit" value={I18n.t("save")} />
+              <input
+                type="submit"
+                value={I18n.t("save")}
+                disabled={this.state.newFiles.length === 0}
+              />
             </div>
           </div>
         </form>